refactor(assets_context): more API
