+++ lib/assets_view.js
@@ -2,7 +2,9 @@
 
 const fs = require('mz/fs');
 const TEMPLATE_CACHE = Symbol('AssetsView#templateCache');
-const HELPER_ASSETS = require('./constant').HELPER_ASSETS;
+const HELPER_ASSETS = require('./util/constant').HELPER_ASSETS;
+const renderDefault = require('./util/default_template');
+
 
 class AssetsView {
   constructor(ctx) {
@@ -21,9 +23,11 @@ class AssetsView {
     const assets = this.ctx.helper.assets;
     assets.setEntry(options.name);
     assets.setContext(options.locals);
-    this.ctx[HELPER_ASSETS] = assets;
 
     if (templateViewEngine && templatePath) {
+      // helper.assets will use this instance
+      this.ctx[HELPER_ASSETS] = assets;
+
       const templateStr = await this.readFileWithCache(templatePath);
       this.logger.info('[egg-view-assets] use %s to render %s, entry is %s',
         templateViewEngine, templatePath, name);
@@ -51,19 +55,3 @@ class AssetsView {
 }
 
 module.exports = AssetsView;
-
-function renderDefault(assets) {
-  return `
-  <!doctype html>
-  <html>
-    <head>
-      ${assets.getStyle()}
-    </head>
-    <body>
-      <div id="root"></div>
-      ${assets.getContext()}
-      ${assets.getScript()}
-    </body>
-  </html>
-  `;
-}
