+++ lib/util/constant.js
